refactor and bugfix
