+++ timesketch/frontend/src/components/Explore/SearchHistoryTree.vue
@@ -5,9 +5,9 @@
 </template>
 
 <script>
-import TreeNode from './TreeNode'
-import ApiClient from '../utils/RestApiClient'
-import EventBus from '../main'
+import TreeNode from './SearchHistoryTreeNode'
+import ApiClient from '../../utils/RestApiClient'
+import EventBus from '../../main'
 
 // Based on https://stackoverflow.com/a/54470906
 function findSearchNode (object, key, predicate) {
