+++ timesketch/frontend/src/components/Explore/SearchHistoryTreeNode.vue
